resize command default

+++ php/common.php
@@ -13639,28 +13639,28 @@ function processFileUploads($docroot=''){
 				elseif(isset($_REQUEST['data-resize']) && strlen($_REQUEST['data-resize'])){
 					$resize=$_REQUEST['data-resize'];
 				}
-				if(strlen($resize)){
+				elseif(isset($CONFIG['resize']) && strlen($CONFIG['resize'])){
+					$resize=$CONFIG['resize'];
+				}
+				if(strlen($resize) && stringContains($file['type'],'image')){
 					if(!isset($CONFIG['resize_command'])){
-						$_REQUEST[$name.'_resize_results']="Error: resize_command not set";	
+						$CONFIG['resize_command']="convert -thumbnail";
 					}
-					elseif(!stringContains($file['type'],'image')){
-						$_REQUEST[$name.'_resize_results']="Error: not an image";
-					}
-					else{
-						$cmd=$CONFIG['resize_command'];
-						$fname=getFileName($abspath,1);
-						$refile=str_replace($fname,$fname.'_resized',$abspath);
+					//resize
+					$cmd=$CONFIG['resize_command'];
+					$fname=getFileName($abspath,1);
+					$refile=str_replace($fname,$fname.'_resized',$abspath);
 
-                		$cmd="{$cmd} {$resize} '{$abspath}' -auto-orient '{$refile}'";
-                		$ok=cmdResults($cmd);
-                		if(is_file($refile) && filesize($refile) > 0){
-							unlink($abspath);
-							rename($refile,$abspath);
-							$_REQUEST[$name.'_size_original']=$_REQUEST[$name.'_size'];
-	                		$_REQUEST[$name.'_size']=filesize($abspath);
-						}
-	                	$_REQUEST[$name.'_resized']=$ok;
-                	}                	
+            		$cmd="{$cmd} {$resize} '{$abspath}' -auto-orient '{$refile}'";
+            		$ok=cmdResults($cmd);
+            		if(is_file($refile) && filesize($refile) > 0){
+						unlink($abspath);
+						rename($refile,$abspath);
+						$_REQUEST[$name.'_size_original']=$_REQUEST[$name.'_size'];
+                		$_REQUEST[$name.'_size']=filesize($abspath);
+					}
+                	$_REQUEST[$name.'_resized']=$ok;
+                                	
 				}
 				//if this is a chunk - see if all chunks are here and combine them.
 				if(isset($_SERVER['HTTP_X_CHUNK_NUMBER']) && isset($_SERVER['HTTP_X_CHUNK_TOTAL'])){
